reorganize stubs in instruments module
